# openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Computer","description":"Computer CRUD"}],"paths":{"/user/{id}":{"get":{"tags":["user-controller"],"operationId":"getUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"put":{"tags":["user-controller"],"operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"delete":{"tags":["user-controller"],"operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/reserve/{id}":{"get":{"tags":["reserve-controller"],"operationId":"getReserve","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Reserve"}}}}}},"put":{"tags":["reserve-controller"],"operationId":"updateReserve","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReserveDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Reserve"}}}}}},"delete":{"tags":["reserve-controller"],"operationId":"deleteReserve","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/game/{id}":{"get":{"tags":["game-controller"],"operationId":"getGame","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Game"}}}}}},"put":{"tags":["game-controller"],"operationId":"updateGame","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Game"}}}}}},"delete":{"tags":["game-controller"],"operationId":"deleteGame","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/computer/{id}":{"get":{"tags":["Computer"],"operationId":"getComputer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Computer"}}}}}},"put":{"tags":["Computer"],"operationId":"updateComputer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComputerDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Computer"}}}}}},"delete":{"tags":["Computer"],"operationId":"deleteComputer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/user/":{"get":{"tags":["user-controller"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}},"post":{"tags":["user-controller"],"operationId":"addUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/reserve/":{"get":{"tags":["reserve-controller"],"operationId":"getAllReserves","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Reserve"}}}}}}},"post":{"tags":["reserve-controller"],"operationId":"addReserve","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReserveDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Reserve"}}}}}}},"/game/":{"get":{"tags":["game-controller"],"operationId":"getAllGames","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Game"}}}}}}},"post":{"tags":["game-controller"],"operationId":"addGame","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Game"}}}}}}},"/computer/":{"get":{"tags":["Computer"],"summary":"Fetch all computers","description":"fetches all computer entities and their data from data source","operationId":"getAllComputers","responses":{"200":{"description":"successful operation","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Computer"}}}}}}},"post":{"tags":["Computer"],"operationId":"addComputer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComputerDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Computer"}}}}}}},"/discount/calculate":{"get":{"tags":["discount-controller"],"operationId":"calculateDiscount","parameters":[{"name":"user","in":"query","required":true,"schema":{"$ref":"#/components/schemas/User"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"number","format":"double"}}}}}}},"/":{"get":{"tags":["main-controller"],"operationId":"getContact_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"ReserveDto":{"required":["computerID","userID"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"userID":{"type":"integer","format":"int64"},"computerID":{"type":"integer","format":"int64"},"time":{"type":"object","properties":{"seconds":{"type":"integer","format":"int64"},"nano":{"type":"integer","format":"int32"},"negative":{"type":"boolean"},"zero":{"type":"boolean"},"units":{"type":"array","items":{"type":"object","properties":{"dateBased":{"type":"boolean"},"timeBased":{"type":"boolean"},"durationEstimated":{"type":"boolean"}}}}}},"price":{"minimum":0,"type":"number","format":"double"}}},"UserDto":{"required":["email","name"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"role":{"type":"string","enum":["ADMIN","CLIENT"]},"email":{"type":"string"},"steamAcc":{"type":"string"},"reservations":{"type":"array","items":{"$ref":"#/components/schemas/ReserveDto"}},"level":{"type":"string","enum":["STARTER","START_LEVEL","MIDDLE_LEVEL","VIP_LEVEL"]},"reservation":{"type":"array","items":{"$ref":"#/components/schemas/ReserveDto"}}}},"Computer":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"price":{"type":"number","format":"double"},"games":{"type":"array","items":{"$ref":"#/components/schemas/Game"}},"reservation":{"type":"array","items":{"$ref":"#/components/schemas/Reserve"}}}},"Game":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"computers":{"type":"array","items":{"$ref":"#/components/schemas/Computer"}}}},"Reserve":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"user":{"$ref":"#/components/schemas/User"},"computer":{"$ref":"#/components/schemas/Computer"},"time":{"type":"object","properties":{"seconds":{"type":"integer","format":"int64"},"nano":{"type":"integer","format":"int32"},"negative":{"type":"boolean"},"zero":{"type":"boolean"},"units":{"type":"array","items":{"type":"object","properties":{"dateBased":{"type":"boolean"},"timeBased":{"type":"boolean"},"durationEstimated":{"type":"boolean"}}}}}},"price":{"type":"number","format":"double"}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"role":{"type":"string","enum":["ADMIN","CLIENT"]},"steamAcc":{"type":"string"},"email":{"type":"string"},"reservations":{"type":"array","items":{"$ref":"#/components/schemas/Reserve"}},"level":{"type":"string","enum":["STARTER","START_LEVEL","MIDDLE_LEVEL","VIP_LEVEL"]},"reservation":{"type":"array","items":{"$ref":"#/components/schemas/Reserve"}}}},"GameDto":{"required":["name"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"ComputerDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"price":{"minimum":0,"type":"number","format":"double"},"reservations":{"type":"array","items":{"$ref":"#/components/schemas/ReserveDto"}},"games":{"type":"array","items":{"$ref":"#/components/schemas/GameDto"}}}}}}}